requirements.txt
# LLM y RAG
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.6

# Embeddings locales
sentence-transformers==2.3.1
torch==2.1.2

# Base de datos vectorial
chromadb==0.4.22

# Procesamiento de documentos
pypdf==4.0.1
python-docx==1.1.0
tiktoken==0.5.2

# Utilidades
python-dotenv==1.0.0
rich==13.7.0
tqdm==4.66.1

# Opcional pero recomendado
numpy==1.26.3